pyproject.toml
[project]
name = "wingman"
version = "0.1.0"
description = "Wingman — personal AI co-worker (LangGraph worker-evaluator loop)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core — Lambda-compatible
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "boto3>=1.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "wikipedia>=1.4.0",
    "bs4>=0.0.2",
    "sendgrid>=6.11.0",
    "reportlab>=4.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
# Install locally for running the dev server and Playwright browser automation
local = [
    "uvicorn[standard]>=0.34.0",
    "playwright>=1.51.0",
    "gradio>=5.22.0",
    "langgraph-checkpoint-sqlite>=2.0.6",
]
